add farsight and yawp, remove better chunk loading
it was causing server crashes, we have other chunk-related mods
anyway so we should be ok without it.

Farsight is VERY important because as of minecraft 1.18+, rendering
is reset and all chunks are refreshed when render distance changes -
serverside mod extends render distance when tick is good and shrinks
when tick is bad which means your chunks constantly refreshing which
is very annoying, farsight fixes the issue.